requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
fastapi==0.128.0
ffmpeg-python @ file:///home/conda/feedstock_root/build_artifacts/ffmpeg-python_1735317926808/work
filelock==3.20.2
fonttools==4.61.1
fsspec==2025.12.0
future @ file:///home/conda/feedstock_root/build_artifacts/future_1738926421307/work
google-auth==2.47.0
google-genai==1.57.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
kiwisolver==1.4.9
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
more-itertools==10.8.0
moviepy==2.2.1
mpmath==1.3.0
networkx==3.6.1
numba @ file:///Users/runner/miniforge3/conda-bld/numba_1765466714240/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1707225620241/work/dist/numpy-1.26.4-cp311-cp311-macosx_10_9_x86_64.whl#sha256=0c694e57489122e5217f1154a05d3d7dd11011371f482a0bcd7692bad5da2ca0
ollama==0.6.1
openai-whisper==20250625
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
polars==1.37.0
polars-runtime-32==1.37.0
proglog==0.1.12
psutil==7.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
scipy==1.17.0
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.252
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.40.0
websockets==15.0.1
